requirements.txt

openai
pytesseract
PyPDF2
transformers
fastapi
uvicorn
python-multipart
Pillow
genai-agentos
